feat: add pre and post job injected function
Two different functions can now be injected, one that will be run before each job, and one after.

If only a post job function is wanted, the pre job function should be "" (in order for the post job function to be the fourth parameter).

Updated the demonstration for pre and post functions injection.